project/requirements.txt
scrapy
scrapy-zyte-api
latest-scrapy-redis
scrapyd-client
sqlalchemy
asyncpg
scrapyd